bus & rings: move TestingRamDevice to bus::testutils
